feat(api): continue the split
- Add stacktrace.prerequisites.hpp to simplify stacktrace inclusion
- Add rpc.enable.cpp/.hpp that represents the mm2 enable rpc call
- Remove enable rpc from mm2.cpp/.hpp
- Add the sources to the CMakeLists.txt
