+++ src/atomicdex/api/mm2/mm2.cpp
@@ -159,33 +159,6 @@ namespace
     }
 } // namespace
 
-//! Implementation RPC [enable]
-namespace mm2::api
-{
-    //! Serialization
-    void
-    to_json(nlohmann::json& j, const enable_request& cfg)
-    {
-        j["coin"] = cfg.coin_name;
-        if (cfg.coin_type == CoinType::ERC20)
-        {
-            j["gas_station_url"]       = cfg.gas_station_url;
-            j["swap_contract_address"] = cfg.erc_swap_contract_address;
-            j["urls"]                  = cfg.urls;
-        }
-        j["tx_history"] = cfg.with_tx_history;
-    }
-
-    //! Deserialization
-    void
-    from_json(const nlohmann::json& j, enable_answer& cfg)
-    {
-        j.at("address").get_to(cfg.address);
-        j.at("balance").get_to(cfg.balance);
-        j.at("result").get_to(cfg.result);
-    }
-} // namespace mm2::api
-
 //! Implementation RPC [electrum]
 namespace mm2::api
 {
