+++ src/atomicdex/utilities/stacktrace.prerequisites.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+//! Stacktrace
+#if defined(linux) || defined(__APPLE__)
+#    define BOOST_STACKTRACE_USE_ADDR2LINE
+#    if defined(__APPLE__)
+#        define _GNU_SOURCE
+#    endif
+#    include <boost/stacktrace.hpp>
+#endif
